move some httpMiddlewares to Middleware
